requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==4.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==4.2.20
django-allauth==65.7.0
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
gspread==6.2.0
idna==3.10
importlib_metadata==8.5.0
l18n==2021.3
laces==0.1.2
Markdown==3.7
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sshtunnel==0.4.0
telepath==0.3.1
typing_extensions==4.13.2
urllib3==2.2.3
wagtail==6.2.3
wagtail-markdown==0.12.1
wagtailcodeblock==1.29.0.2
webencodings==0.5.1
Willow==1.8.0
zipp==3.20.2